Add ManageIQ::Providers::CloudManager::OrchestrationStack as an intermediate base
