+++ app/models/manageiq/providers/openstack/infra_manager/orchestration_stack.rb
@@ -1,5 +1,20 @@
 class ManageIQ::Providers::Openstack::InfraManager::OrchestrationStack < ::OrchestrationStack
   belongs_to :ext_management_system, :foreign_key => :ems_id, :class_name => "ManageIQ::Providers::InfraManager"
+  belongs_to :orchestration_template
+  belongs_to :cloud_tenant
+
+  def self.create_stack(orchestration_manager, stack_name, template, options = {})
+    klass = orchestration_manager.class::OrchestrationStack
+    ems_ref = klass.raw_create_stack(orchestration_manager, stack_name, template, options)
+
+    klass.create(:name                   => stack_name,
+                 :ems_ref                => ems_ref,
+                 :status                 => 'CREATE_IN_PROGRESS',
+                 :resource_group         => options[:resource_group],
+                 :ext_management_system  => orchestration_manager,
+                 :cloud_tenant           => tenant,
+                 :orchestration_template => template)
+  end
 
   def raw_update_stack(template, parameters)
     ext_management_system.with_provider_connection(:service => "Orchestration") do |connection|
