+++ app/models/service_orchestration.rb
@@ -49,8 +49,8 @@ class ServiceOrchestration < Service
   end
 
   def deploy_orchestration_stack
-    @orchestration_stack =
-      OrchestrationStack.create_stack(orchestration_manager, stack_name, orchestration_template, stack_options)
+    @orchestration_stack = ManageIQ::Providers::CloudManager::OrchestrationStack.create_stack(
+      orchestration_manager, stack_name, orchestration_template, stack_options)
     add_resource(@orchestration_stack)
     @orchestration_stack
   ensure
