+++ lib/miq_automation_engine/service_models/miq_ae_service_manageiq-providers-azure-cloud_manager-orchestration_stack.rb
@@ -1,4 +1,4 @@
 module MiqAeMethodService
-  class MiqAeServiceManageIQ_Providers_Azure_CloudManager_OrchestrationStack < MiqAeServiceOrchestrationStack
+  class MiqAeServiceManageIQ_Providers_Azure_CloudManager_OrchestrationStack < MiqAeServiceManageIQ_Providers_CloudManager_OrchestrationStack
   end
 end
