+++ lib/miq_automation_engine/service_models/miq_ae_service_manageiq-providers-cloud_manager-orchestration_stack.rb
@@ -1,0 +1,8 @@
+module MiqAeMethodService
+  class MiqAeServiceManageIQ_Providers_CloudManager_OrchestrationStack < MiqAeServiceOrchestrationStack
+    expose :vms,                    :association => true
+    expose :security_groups,        :association => true
+    expose :cloud_networks,         :association => true
+    expose :orchestration_template, :association => true
+  end
+end
