+++ lib/miq_automation_engine/service_models/miq_ae_service_manageiq-providers-openstack-cloud_manager-orchestration_stack.rb
@@ -1,3 +1,4 @@
 module MiqAeMethodService
-  class MiqAeServiceManageIQ_Providers_Openstack_CloudManager_OrchestrationStack < MiqAeServiceOrchestrationStack; end
+  class MiqAeServiceManageIQ_Providers_Openstack_CloudManager_OrchestrationStack < MiqAeServiceManageIQ_Providers_CloudManager_OrchestrationStack
+  end
 end
