+++ spec/models/manageiq/providers/azure/cloud_manager/orchestration_stack_spec.rb
@@ -28,7 +28,7 @@ describe ManageIQ::Providers::Azure::CloudManager::OrchestrationStack do
       it 'creates a stack' do
         expect(orchestration_service).to receive(:create).and_return(the_raw_stack)
 
-        stack = OrchestrationStack.create_stack(ems, 'mystack', template, {})
+        stack = ManageIQ::Providers::CloudManager::OrchestrationStack.create_stack(ems, 'mystack', template, {})
         expect(stack.class).to   eq(described_class)
         expect(stack.name).to    eq('mystack')
         expect(stack.ems_ref).to eq(the_raw_stack.id)
@@ -37,7 +37,9 @@ describe ManageIQ::Providers::Azure::CloudManager::OrchestrationStack do
       it 'catches errors from provider' do
         expect(orchestration_service).to receive(:create).and_throw('bad request')
 
-        expect { OrchestrationStack.create_stack(ems, 'mystack', template, {}) }.to raise_error(MiqException::MiqOrchestrationProvisionError)
+        expect do
+          ManageIQ::Providers::CloudManager::OrchestrationStack.create_stack(ems, 'mystack', template, {})
+        end.to raise_error(MiqException::MiqOrchestrationProvisionError)
       end
     end
